delete matchNames.ts
