--- src/renderer/lib/matchNames.ts
@@ -1,24 +1,0 @@
-export function namesMatch(lookingForNametags: string[], inGameTags: string[], fuzzyMatch = true): boolean {
-  if (lookingForNametags.length === 0 || inGameTags.length === 0) {
-    return false;
-  }
-
-  const match = inGameTags.find((name) => {
-    // If we're not doing fuzzy matching just return the exact match
-    if (!fuzzyMatch) {
-      return lookingForNametags.includes(name);
-    }
-
-    // Replace the netplay names with underscores and coerce to lowercase
-    // Smashladder internally represents spaces as underscores when writing SLP files
-    const fuzzyNetplayName = name.toLowerCase();
-    const matchedFuzzyTag = lookingForNametags.find((tag) => {
-      const lowerSearch = tag.toLowerCase();
-      const fuzzySearch = tag.split(" ").join("_").toLowerCase();
-      return fuzzyNetplayName.startsWith(lowerSearch) || fuzzyNetplayName.startsWith(fuzzySearch);
-    });
-    return matchedFuzzyTag !== undefined;
-  });
-
-  return match !== undefined;
-}
