+++ src/renderer/lib/replayFileSort.ts
@@ -3,8 +3,6 @@ import type { FileResult } from "@replays/types";
 import { Frames, GameMode } from "@slippi/slippi-js";
 import compareFunc from "compare-func";
 
-import { namesMatch } from "@/lib/matchNames";
-
 // The minimum duration of games when filtering out short games
 const MIN_GAME_DURATION_FRAMES = 30 * 60;
 const STADIUM_GAME_MODES = [GameMode.HOME_RUN_CONTEST, GameMode.TARGET_TEST];
@@ -110,4 +108,29 @@ function matchablePlayerNames(file: FileResult): string[] {
   return file.game.players.flatMap((p) => {
     return [p.displayName, p.tag, p.connectCode].filter(exists);
   });
+}
+
+function namesMatch(lookingForNametags: string[], inGameTags: string[], fuzzyMatch = true): boolean {
+  if (lookingForNametags.length === 0 || inGameTags.length === 0) {
+    return false;
+  }
+
+  const match = inGameTags.find((name) => {
+    // If we're not doing fuzzy matching just return the exact match
+    if (!fuzzyMatch) {
+      return lookingForNametags.includes(name);
+    }
+
+    // Replace the netplay names with underscores and coerce to lowercase
+    // Smashladder internally represents spaces as underscores when writing SLP files
+    const fuzzyNetplayName = name.toLowerCase();
+    const matchedFuzzyTag = lookingForNametags.find((tag) => {
+      const lowerSearch = tag.toLowerCase();
+      const fuzzySearch = tag.split(" ").join("_").toLowerCase();
+      return fuzzyNetplayName.startsWith(lowerSearch) || fuzzyNetplayName.startsWith(fuzzySearch);
+    });
+    return matchedFuzzyTag !== undefined;
+  });
+
+  return match !== undefined;
 }
